Move files to appointments folder
